+++ examples/twitter-clone/src/main/scala/finatra/quickstart/domain/http/TweetGetRequest.scala
@@ -1,6 +1,6 @@
 package finatra.quickstart.domain.http
 
-import com.twitter.finatra.request.RouteParam
+import com.twitter.finatra.http.annotations.RouteParam
 import finatra.quickstart.domain.TweetId
 
 case class TweetGetRequest(@RouteParam id: TweetId)