+++ http-annotations/src/main/java/com/twitter/finatra/http/annotations/QueryParam.java
@@ -1,17 +1,30 @@
-package com.twitter.finatra.request;
+package com.twitter.finatra.http.annotations;
 
+import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 import com.twitter.finatra.json.annotations.InjectableValue;
 
-import static java.lang.annotation.ElementType.PARAMETER;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-@Target({ PARAMETER })
-@Retention(RUNTIME)
+/**
+ * Marker {@link java.lang.annotation.Annotation} for denoting a Jackson "injectable value" which
+ * should be obtained from a Finagle HTTP Request query string parameter.
+ *
+ * @see com.twitter.finatra.json.annotations.InjectableValue
+ * @see <a href="https://twitter.github.io/finatra/user-guide/json/index.html#injectablevalues">Finatra User's Guide - JSON Injectable Values</a>
+ * @see <a href="https://twitter.github.io/finatra/user-guide/http/requests.html#field-annotations">Finatra User's Guide - HTTP Request Field Annotations</a>
+ */
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
 @InjectableValue
 public @interface QueryParam {
+    /**
+     * An optional field name to use for reading the query parameter from the Finagle HTTP Request.
+     * When empty, the annotated case class field name will be used to read the query parameter
+     * from the Finagle HTTP Request parameters.
+     * @return the name of the query parameter field.
+     */
     String value() default "";
 
     /**