+++ http/src/test/scala/com/twitter/finatra/http/tests/integration/doeverything/main/domain/UserAgentRequest.scala
@@ -1,5 +1,5 @@
 package com.twitter.finatra.http.tests.integration.doeverything.main.domain
 
-import com.twitter.finatra.request.Header
+import com.twitter.finatra.http.annotations.Header
 
 case class UserAgentRequest(@Header("user-agent") agent: String)