+++ http/src/test/scala/com/twitter/finatra/http/tests/integration/tweetexample/main/controllers/AdminController.scala
@@ -1,9 +1,8 @@
 package com.twitter.finatra.http.tests.integration.tweetexample.main.controllers
 
 import com.twitter.finagle.http.Request
-import com.twitter.finatra.http.Controller
+import com.twitter.finatra.http.{Controller, Prod, Staging}
 import com.twitter.finatra.http.tests.integration.tweetexample.main.services.admin.UserService
-import com.twitter.finatra.test.{Prod, Staging}
 import javax.inject.Inject
 
 class AdminController @Inject()(@Prod prodUsers: UserService, @Staging stagingUsers: UserService)