+++ http/src/test/scala/com/twitter/finatra/http/tests/marshalling/MessageBodyManagerTest.scala
@@ -2,7 +2,8 @@ package com.twitter.finatra.http.tests.marshalling
 
 import com.fasterxml.jackson.databind.JsonNode
 import com.twitter.finagle.http.{Fields, MediaType, Message, Request, Response}
-import com.twitter.finatra.http.TestMessageBodyWriterAnn
+import com.twitter.finatra.http.{Prod, TestMessageBodyWriterAnn}
+import com.twitter.finatra.http.annotations.{Header, QueryParam}
 import com.twitter.finatra.http.marshalling._
 import com.twitter.finatra.http.modules.MessageBodyModule
 import com.twitter.finatra.http.tests.integration.json.CaseClassWithBoolean
@@ -10,8 +11,6 @@ import com.twitter.finatra.jackson.ScalaObjectMapper
 import com.twitter.finatra.jackson.caseclass.exceptions.InjectableValuesException
 import com.twitter.finatra.jackson.modules.ScalaObjectMapperModule
 import com.twitter.finatra.modules.FileResolverModule
-import com.twitter.finatra.request.{Header, QueryParam}
-import com.twitter.finatra.test.Prod
 import com.twitter.inject.app.TestInjector
 import com.twitter.inject.{Injector, Mockito, Test}
 import javax.inject.Inject