+++ jackson/src/test/scala/com/twitter/finatra/jackson/tests/serde/JodaDatetimeDeserializerTest.scala
@@ -3,7 +3,9 @@ package com.twitter.finatra.jackson.tests.serde
 import com.fasterxml.jackson.databind.{SerializationFeature, ObjectMapper => JacksonObjectMapper}
 import com.fasterxml.jackson.datatype.joda.JodaModule
 import com.fasterxml.jackson.module.scala.DefaultScalaModule
-import com.fasterxml.jackson.module.scala.experimental.{ScalaObjectMapper => JacksonScalaObjectMapper}
+import com.fasterxml.jackson.module.scala.experimental.{
+  ScalaObjectMapper => JacksonScalaObjectMapper
+}
 import com.twitter.finatra.jackson.ScalaObjectMapper
 import com.twitter.finatra.jackson.serde.SerDeSimpleModule
 import com.twitter.inject.Test